fixed info AlertDialog
