+++ app/src/main/res/layout/info_dialog_view.xml
@@ -11,5 +11,7 @@
         android:layout_alignParentTop="true"
         android:layout_alignParentLeft="true"
         android:layout_alignParentStart="true"
-        android:checked="false" />
+        android:checked="false"
+        android:textColor="#000000"
+        android:layout_marginTop="50sp" />
 </RelativeLayout>
